pyproject.toml
[project]
name = "minillm"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "haruki.sato01", email = "haruki.sato01@g.softbank.co.jp" }
]
dependencies = [
    "lightning>=2.2.0",
    "deepspeed>=0.13.1",
    "datasets>=2.17.0",
    "torch-summary>=1.4.5",
    "pandas>=2.0.3",
    "ipynb>=0.5.1",
    "notebook>=7.0.8",
    "torchtext>=0.16.2",
    "torchvision>=0.17.0",
    "numpy>=1.24.4",
    "janome>=0.5.0",
    "spacy>=3.7.2",
    "xlrd>=2.0.1",
    "pip>=24.0",
    "black>=24.1.1",
    "isort>=5.13.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "minillm:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/minillm"]
